% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Returns a custom dataset of a requested set of indicators for all available
hospitals over a given time frame.}
\usage{
get_datasets(indicator_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{indicator_id}{The ID of an indicator to show data for. The ID for a given
indicator can be obtained from the indicators endpoint. More than one
indicator ID can be specified}

\item{start_date, end_date}{(character or `Date` object) Specifies that only
data _after_ (for start) and/or _before_ (for end) a certain point should
be returned. The value can be a date of the form `YYYY`, `YYYY-MM` or
`YYYY-MM-DD` (`Date` objects will be converted to the long ISO form)
Where only a year is specified, the search period starts at the beginning
of the year, where a year and month are specified, the search period
starts at the beginning of the month. Both default to `NULL` as they
are optional.}
}
\value{
list with dataset metadata and associated indicator measures
}
\description{
Given the unlikely possibility of this data changing in-session, the
result is memoised. To save the API bandwidth/CPU.
}
\references{
<https://www.myhospitals.gov.au/about-the-data/download-data>
}
