% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-hospitals.R
\name{get_hospitals}
\alias{get_hospitals}
\title{Returns a data frame of all (public) hospitals with data in the MyHospitals website.}
\usage{
get_hospitals()
}
\value{
data frame (tibble)
}
\description{
Given the unlikely possibility of this data changing in-session, the
result is memoised. To save the API bandwidth/CPU.
}
\references{
<https://www.myhospitals.gov.au/about-the-data/download-data>
}
