% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Returns a data frame of all indicators measured within the MyHospitals website.}
\usage{
get_indicators()
}
\value{
data frame (tibble)
}
\description{
Given the unlikely possibility of this data changing in-session, the
result is memoised. To save the API bandwidth/CPU.
}
\note{
Not all hospitals will have a data stored for an indicator.
}
\references{
<https://www.myhospitals.gov.au/about-the-data/download-data>
}
