% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosperf-package.R
\docType{package}
\name{hosperf}
\alias{hosperf}
\alias{hosperf-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/hosperf}
\item BugReports: \url{https://gitlab.com/hrbrmstr/hosperf/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
